Fix version future-proofing with is_uma_lnurlp_query.
Future versions of UMA may change fields in this request, etc. so treating a parse failure as "non-uma" isn't the right behavior. We should allow the parsing to spit up the proper UnsupportedVersion exception later. This was causing the demo vasp to treat v1 requests as regular lnurl, rather than negotiating a lower UMA version.